Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/2699



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@11591 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/strprep.h
@@ -316,7 +316,7 @@ inline UBool StringPrep::getNormalization(){
 }
 
 inline void StringPrep::getUnicodeVersion(UVersionInfo& info){
-    for(int32_t i=0; i< (sizeof(info)/sizeof(info[0])); i++){
+    for(int32_t i=0; i< (int32_t)(sizeof(info)/sizeof(info[0])); i++){
         info[i] = unicodeVersion[i];
     }
 }
